requirements.txt
environ
Django==3.0.7
gmaps==0.9.0
gmplot==1.4.0
gunicorn==20.0.4
psycopg2==2.8.5
dj-static==0.0.6
django-crispy-forms
biopython
django-environ